pyproject.toml
[tool.poetry]
name = "cloud-storage"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
packages = [{include = "src"}]

#[tool.poetry.scripts]
#start = "manage:run"

[tool.poetry.dependencies]
python = "^3.11"
docker = "^7.1.0"
django = "^5.1.7"
djangorestframework = "^3.15.2"
redis = "^5.2.1"
django-redis = "^5.4.0"
python-dotenv = "^1.0.1"
drf-yasg = "^1.21.10"
black = "^25.1.0"
psycopg2-binary = "^2.9.10"
django-htmx = "^1.23.0"
pytest-django = "^4.10.0"
factory-boy = "^3.3.3"
boto3 = "^1.37.13"
django-storages = "^1.14.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
